Rust 1.28.0-dev

Rust Documentation

Welcome to an overview of the documentation provided by the Rust project. All of these projects are managed by the Docs Team; there are other unofficial documentation resources as well!

Many of these resources take the form of "books"; we collectively call these "The Rust Bookshelf." Some are large, some are small.

1 Learn Rust

If you'd like to learn Rust, this is the spot for you! All of these resources assume that you have programmed before, but not in any specific language:

1.1 The Rust Programming Language

Affectionately nicknamed "the book," The Rust Programming Language will give you an overview of the language from first principles. You'll build a few projects along the way, and by the end, you'll have a solid grasp of the language.

1.2 Rust By Example

If reading multiple hundreds of pages about a language isn't your style, then Rust By Example has you covered. While the book talks about code with a lot of words, RBE shows off a bunch of code, and keeps the talking to a minimum. It also includes exercises!

2 Use Rust

Once you've gotten familliar with the language, these resources can help you when you're actually using it day-to-day.

2.1 The Standard Library

Rust's standard library has extensive API documentation, with explanations of how to use various things, as well as example code for accomplishing various tasks.

2.2 The Rustc Book

The Rustc Book describes the Rust compiler, rustc.

2.3 The Cargo Book

The Cargo Book is a guide to Cargo, Rust's build tool and dependency manager.

2.4 The Rustdoc Book

The Rustdoc Book describes our documentation tool, rustdoc.

2.5 Extended Error Listing

Many of Rust's errors come with error codes, and you can request extended diagnostics from the compiler on those errors. You can also read them here, if you prefer to read them that way.

3 Master Rust

Once you're quite familiar with the language, you may find these advanced resources useful.

3.1 The Reference

The Reference is not a formal spec, but is more detailed and comprehensive than the book.

3.2 The Rustonomicon

The Rustonomicon is your guidebook to the dark arts of unsafe Rust. It's also sometimes called "the 'nomicon."

3.3 The Unstable Book

The Unstable Book has documentation for unstable features.