Rust 1.28.0-dev

The Rust Programming Language

“The Rust Programming Language,” is an introductory book about Rust, with multiple editions.

The source for all editions lives on GitHub. Please open issues with any questions, concerns, or tweaks.

2018 Edition

The 2018 edition of the book is a "living version" of the book; based on the second edition, it will be updated as Rust updates. If you're not sure what version of the book to read, you should prefer this version.

You can find the 2018 edition here.

Second Edition

The second edition is a complete re-write of TRPL, from the ground up, and is therefore very different from the first edition. It's also based on Rust 1.21, and so may be missing features that landed in stable after that release.

You can purchase a dead-tree copy from NoStarch Press.

This edition of the book is frozen; it contains exactly what's in the printed book.

You can find the second edition here.

First Edition

The first edition is completely frozen, but is still included in the documentation for historical reasons.

You can find the first edition here.